// frontend/src/lib/api/generated/models/ListingEventPurchasedTicketsItem.ts
/**
 * Roominar
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export class ListingEventPurchasedTicketsItem {
    'transactionItemId': number;
    'checkedInAt'?: Date | null;
    'checkInId'?: number | null;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "transactionItemId",
            "baseName": "transaction_item_id",
            "type": "number",
            "format": ""
        },
        {
            "name": "checkedInAt",
            "baseName": "checked_in_at",
            "type": "Date",
            "format": "date-time"
        },
        {
            "name": "checkInId",
            "baseName": "check_in_id",
            "type": "number",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return ListingEventPurchasedTicketsItem.attributeTypeMap;
    }

    public constructor() {
    }
}